// src/app/interfaces/weather.ts
export interface IWeather {
  cities: IWeather;
  base: string;
  clouds?: {
    all: 2;
  };
  cod: number;
  coord: {
    lon: number;
    lat: number;
  };
  dt: number;
  main: {
    feels_like: number;
    humidity: number;
    pressure: number;
    temp: number;
    temp_max: number;
    temp_min: number;
  };
  name: string;
  sys: {
    country: string;
    id: number;
    sunrise: number;
    sunset: number;
    type: number;
  };
  timezone: number;
  visibility: number;
  weather: Array<{
    description: string;
    icon: string;
    id: number;
    main: string;
  }>;
  wind?: {
    speed: number;
    deg: number;
  };
}